LayoutTests/fast/url/query-expected.txt

Test URLs that have a query string.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS canonicalize('http://www.example.com/?foo=bar') is 'http://www.example.com/?foo=bar'
PASS canonicalize('http://www.example.com/?as?df') is 'http://www.example.com/?as?df'
FAIL canonicalize('http://www.example.com/?as#df') should be http://www.example.com/?as%23df. Was http://www.example.com/?as#df.
FAIL canonicalize('http://www.example.com/?hello bye') should be http://www.example.com/?%02hello%7F%20bye. Was http://www.example.com/?hello%7F%20bye.
PASS canonicalize('http://www.example.com/?%40%41123') is 'http://www.example.com/?%40%41123'
PASS canonicalize('http://www.example.com/?q=ä½ å¥½') is 'http://www.example.com/?q=%E4%BD%A0%E5%A5%BD'
FAIL canonicalize('http://www.example.com/?q=íí') should be http://www.example.com/?q=%EF%BF%BD%EF%BF%BD. Was http://www.example.com/?q=%ED%ED.
PASS canonicalize('http://www.example.com/?q=<asdf>') is 'http://www.example.com/?q=%3Casdf%3E'
PASS canonicalize('http://www.example.com/?q="asdf"') is 'http://www.example.com/?q=%22asdf%22'
PASS successfullyParsed is true

TEST COMPLETE
